Read OK for team, getAll Official

+++ services/officialLeague/src/main/java/ch/heigvd/amt/repositories/GameRepository.java
@@ -3,5 +3,5 @@ package ch.heigvd.amt.repositories;
 import ch.heigvd.amt.entities.GameEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface GameRepository extends CrudRepository<GameEntity, Long> {
+public interface GameRepository extends CrudRepository<GameEntity, Integer> {
 }
