+++ services/officialLeague/src/main/java/ch/heigvd/amt/repositories/OfficialRepository.java
@@ -3,5 +3,5 @@ package ch.heigvd.amt.repositories;
 import ch.heigvd.amt.entities.OfficialEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface OfficialRepository extends CrudRepository<OfficialEntity, Long> {
+public interface OfficialRepository extends CrudRepository<OfficialEntity, Integer> {
 }
