+++ services/officialLeague/src/main/java/ch/heigvd/amt/repositories/TeamRepository.java
@@ -3,6 +3,6 @@ package ch.heigvd.amt.repositories;
 import ch.heigvd.amt.entities.TeamEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TeamRepository extends CrudRepository<TeamEntity, Long> {
+public interface TeamRepository extends CrudRepository<TeamEntity, Integer> {
     
 }
